+++ src/mantle/facade/class-log.php
@@ -7,29 +7,32 @@
 
 namespace Mantle\Facade;
 
+use DateTimeImmutable;
+use Mantle\Support\Stringable;
+
 /**
  * Log Facade
  *
  * @method static \Mantle\Log\Logger set_dispatcher(\Mantle\Contracts\Events\Dispatcher $dispatcher = null)
  * @method static bool addRecord(int $level, string $message, mixed[] $context = [], DateTimeImmutable $datetime = null)
  * @method static string getName()
- * @method static \self withName(string $name)
- * @method static \self pushHandler(\Monolog\Handler\HandlerInterface $handler)
+ * @method static \Mantle\Log\Logger withName(string $name)
+ * @method static \Mantle\Log\Logger pushHandler(\Monolog\Handler\HandlerInterface $handler)
  * @method static \Monolog\Handler\HandlerInterface popHandler()
- * @method static \self setHandlers(\Monolog\Handler\HandlerInterface[] $handlers)
+ * @method static \Mantle\Log\Logger setHandlers(\Monolog\Handler\HandlerInterface[] $handlers)
  * @method static \Monolog\Handler\HandlerInterface[] getHandlers()
- * @method static \self pushProcessor(callable $callback)
+ * @method static \Mantle\Log\Logger pushProcessor(callable $callback)
  * @method static callable popProcessor()
  * @method static callable[] getProcessors()
- * @method static \self useMicrosecondTimestamps(bool $micro)
- * @method static \self useLoggingLoopDetection(bool $detectCycles)
+ * @method static \Mantle\Log\Logger useMicrosecondTimestamps(bool $micro)
+ * @method static \Mantle\Log\Logger useLoggingLoopDetection(bool $detectCycles)
  * @method static void close()
  * @method static void reset()
  * @method static array getLevels()
  * @method static string getLevelName(int $level)
  * @method static int toMonologLevel(string|int $level)
  * @method static bool isHandling(int $level)
- * @method static \self setExceptionHandler(callable|null $callback)
+ * @method static \Mantle\Log\Logger setExceptionHandler(callable|null $callback)
  * @method static callable|null getExceptionHandler()
  * @method static void log(mixed $level, string|Stringable $message, mixed[] $context = [])
  * @method static void debug(string|Stringable $message, mixed[] $context = [])
@@ -40,7 +43,7 @@ namespace Mantle\Facade;
  * @method static void critical(string|Stringable $message, mixed[] $context = [])
  * @method static void alert(string|Stringable $message, mixed[] $context = [])
  * @method static void emergency(string|Stringable $message, mixed[] $context = [])
- * @method static \self setTimezone(\DateTimeZone $tz)
+ * @method static \Mantle\Log\Logger setTimezone(\DateTimeZone $tz)
  * @method static \DateTimeZone getTimezone()
  *
  * @see \Mantle\Log\Logger