+++ src/mantle/support/interface-enumerable.php
@@ -45,7 +45,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  callable|null $callback
 	 * @return static
 	 */
-	public static function times( $number, callable $callback = null );
+	public static function times( $number, ?callable $callback = null );
 
 	/**
 	 * Wrap the given value in a collection if applicable.
@@ -265,7 +265,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable(TValue): bool)|null  $callback
 	 * @return static
 	 */
-	public function filter( callable $callback = null );
+	public function filter( ?callable $callback = null );
 
 	/**
 	 * Apply the callback if the value is truthy.
@@ -278,7 +278,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable($this, TWhenParameter): TWhenReturnType)|null  $default
 	 * @return static|TWhenReturnType
 	 */
-	public function when( $value, callable $callback = null, callable $default = null );
+	public function when( $value, ?callable $callback = null, ?callable $default = null );
 
 	/**
 	 * Apply the callback if the collection is empty.
@@ -289,7 +289,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable($this): TWhenEmptyReturnType)|null  $default
 	 * @return $this|TWhenEmptyReturnType
 	 */
-	public function when_empty( callable $callback, callable $default = null );
+	public function when_empty( callable $callback, ?callable $default = null );
 
 	/**
 	 * Apply the callback if the collection is not empty.
@@ -300,7 +300,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable($this): TWhenEmptyReturnType)|null  $default
 	 * @return $this|TWhenEmptyReturnType
 	 */
-	public function when_not_empty( callable $callback, callable $default = null );
+	public function when_not_empty( callable $callback, ?callable $default = null );
 
 	/**
 	 * Apply the callback if the value is falsy.
@@ -313,7 +313,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable( $this, TUnlessParameter): TUnlessReturnType)|null  $default
 	 * @return $this|TUnlessReturnType
 	 */
-	public function unless( $value, callable $callback = null, callable $default = null );
+	public function unless( $value, ?callable $callback = null, ?callable $default = null );
 
 	/**
 	 * Apply the callback unless the collection is empty.
@@ -324,7 +324,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable($this): TUnlessEmptyReturnType)|null  $default
 	 * @return $this|TUnlessEmptyReturnType
 	 */
-	public function unless_empty( callable $callback, callable $default = null );
+	public function unless_empty( callable $callback, ?callable $default = null );
 
 	/**
 	 * Apply the callback unless the collection is not empty.
@@ -335,7 +335,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  (callable($this): TUnlessNotEmptyReturnType)|null  $default
 	 * @return $this|TUnlessNotEmptyReturnType
 	 */
-	public function unless_not_empty( callable $callback, callable $default = null );
+	public function unless_not_empty( callable $callback, ?callable $default = null );
 
 	/**
 	 * Filter items by the given key value pair.
@@ -431,7 +431,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  TFirstDefault|(\Closure(): TFirstDefault)  $default
 	 * @return TValue|TFirstDefault
 	 */
-	public function first( callable $callback = null, $default = null );
+	public function first( ?callable $callback = null, $default = null );
 
 	/**
 	 * Get the first item by the given key value pair.
@@ -546,7 +546,7 @@ interface Enumerable extends Arrayable, Countable, IteratorAggregate, Jsonable, 
 	 * @param  TLastDefault|(\Closure(): TLastDefault)  $default
 	 * @return TValue|TLastDefault
 	 */
-	public function last( callable $callback = null, $default = null );
+	public function last( ?callable $callback = null, $default = null );
 
 	/**
 	 * Run a map over each of the items.